TON-1406 Allow to pass custom fetch manifest function into config

+++ Sources/TONWalletKit/API/Models/WalletKit/TONSwapQuoteParams.swift
@@ -34,15 +34,15 @@ public struct TONSwapQuoteParams<TProviderOptions: Codable>: Codable {
     public var to: TONSwapToken
     public var network: TONNetwork
     /** Slippage tolerance in basis points (1 bp = 0.01%) */
-    public var slippageBps: Double?
+    public var slippageBps: Int?
     /** Maximum number of outgoing messages */
-    public var maxOutgoingMessages: Double?
+    public var maxOutgoingMessages: Int?
     /** Provider-specific options */
     public var providerOptions: TProviderOptions?
     /** If true, amount is the amount to receive (buy). If false, amount is the amount to spend (sell). */
     public var isReverseSwap: Bool?
 
-    public init(amount: String, from: TONSwapToken, to: TONSwapToken, network: TONNetwork, slippageBps: Double? = nil, maxOutgoingMessages: Double? = nil, providerOptions: TProviderOptions? = nil, isReverseSwap: Bool? = nil) {
+    public init(amount: String, from: TONSwapToken, to: TONSwapToken, network: TONNetwork, slippageBps: Int? = nil, maxOutgoingMessages: Int? = nil, providerOptions: TProviderOptions? = nil, isReverseSwap: Bool? = nil) {
         self.amount = amount
         self.from = from
         self.to = to
